MMC: fix hang if card was removed during suspend and unsafe resume was enabled
Currently removal of the card leads to del_disk called indirectly by mmc core.
This function expects userspace to be running, which isn't when .resume is called

Fix that by removing the code that did that in mmc_resume_host. It is possible
because card detection logic will kick it later and remove the card.

Also make mtd workqueue freezeable, so it won't attempt to add/remove the card
while userspace is frozen.

Signed-off-by: Maxim Levitsky <maximlevitsky@gmail.com>